feat: Add draft architecture
This is not at all fixed. Just an idea of how we might go about
implementing the assignment. It's rather simple and allows for easy
replacement of different weather sources and UI views for assignment2.
Using repository pattern for data. MVP for easy testing and separation
of views and data.